Updated game model, abstract zone, controller package. Added call back interface. Added instantiation to MainApp, with useful notes to use.

+++ src/main/java/model/common/Satisfaction.java
@@ -61,7 +61,7 @@ public class Satisfaction {
         this.taxEffect = taxEffect;
     }
 
-    public void setZoneBalanceEffect(int indComZoneBalance) {
+    public void setZoneBalanceEffect(float indComZoneBalance) {
         this.indComZoneBalance = indComZoneBalance;
     }
